Resolve non-conflicting merge problems
